requirements.txt
fastapi==0.115.0
mangum==0.18.0
boto3==1.35.0
pydantic==2.9.0
pydantic-settings==2.5.2
python-multipart==0.0.12
Pillow==10.4.0